Inform the user on search results
Let the user know number of results displayed and found. GitHub shows 30 users per page and don't want to page through all to build a filter list yet. See if people raise this as an issue.

Not guarenteed an email from api so will throw an error message if someone picks an item without an email.

#15